music/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "hibiki-music"
version = "0.1.0"
description = "智能原生 macOS 音乐播放器 - 基于 Hibiki UI"
authors = [{ name = "Hibiki Music Team" }]
dependencies = [
    # Hibiki UI 框架依赖
    "hibiki-ui",
    # 音频处理
    "mutagen>=1.47.0",
    "pyobjc-framework-AVFoundation>=10.0",
    "pyobjc-framework-MediaPlayer>=10.0",
    # 数据库 - 现代化 SQLModel 架构
    "sqlmodel>=0.0.14",
    "sqlalchemy>=2.0.0",
    "alembic>=1.12.0",
    # AI/NLP 功能
    "langdetect>=1.0.9",
    "scikit-learn>=1.3.0",
    "jieba>=0.42.1",
    "opencc-python-reimplemented>=0.1.7",
    # 工具库
    "pydantic>=2.4.0",
    "python-dateutil>=2.8.0",
    "pathlib2>=2.3.0",
    "watchdog>=3.0.0",
    # 开发工具
    "pytest>=7.4.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.5.0",
    "pillow>=11.3.0",
    "cairosvg>=2.8.2",
]
readme = "README.md"
requires-python = ">= 3.11"

[project.scripts]
music = 'hibiki.music.main:main'

[tool.hatch.build.targets.wheel]
packages = ["src/hibiki"]

[tool.uv.sources]
hibiki-ui = { workspace = true }

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
